+++ provider/pro/resource_rediscloud_pro_subscription.go
@@ -631,30 +631,9 @@ func resourceRedisCloudProSubscriptionCreate(ctx context.Context, d *schema.Reso
 		return append(diags, diag.FromErr(err)...)
 	}
 
-	// There is a timing issue where the subscription is marked as active before the creation-plan databases are listed.
-	// This additional wait ensures that the databases will be listed before calling api.client.Database.List()
-	time.Sleep(30 * time.Second) //lintignore:R018
-	if err := utils.WaitForSubscriptionToBeActive(ctx, subId, api); err != nil {
-		return append(diags, diag.FromErr(err)...)
-	}
-
-	// Locate Databases to confirm Active status
-	dbList := api.Client.Database.List(ctx, subId)
-
-	for dbList.Next() {
-		dbId := *dbList.Value().ID
-
-		if err := utils.WaitForDatabaseToBeActive(ctx, subId, dbId, api); err != nil {
-			return append(diags, diag.FromErr(err)...)
-		}
-		// Delete each creation-plan database
-		dbErr := api.Client.Database.Delete(ctx, subId, dbId)
-		if dbErr != nil {
-			diag.FromErr(dbErr)
-		}
-	}
-	if dbList.Err() != nil {
-		return append(diags, diag.FromErr(dbList.Err())...)
+	// Delete databases created by the subscription creation plan
+	if cleanupDiags := utils.DeleteSubscriptionDatabases(ctx, subId, api); cleanupDiags != nil {
+		return append(diags, cleanupDiags...)
 	}
 
 	if redisVersion != "" {
@@ -913,6 +892,11 @@ func resourceRedisCloudProSubscriptionUpdateCmk(ctx context.Context, d *schema.R
 
 	if err := utils.WaitForSubscriptionToBeActive(ctx, subId, api); err != nil {
 		return diag.FromErr(err)
+	}
+
+	// After CMK activation, delete databases that were not cleaned up during Create
+	if cleanupDiags := utils.DeleteSubscriptionDatabases(ctx, subId, api); cleanupDiags != nil {
+		return cleanupDiags
 	}
 
 	return nil